refactor: Application에서 Controller 분리, ChessController의 @Controller와 @RestController 분리
